go/worker: Move RuntimeHostNode into the common node
